+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.8.2] - 2025-02-10
 ### Fixed
 - Added missing flushing limit values, and setup automated flush if limit is reached
+### Changed
+- Change LLM tracking latency to seconds and remove default latency from event properties
 
 
 ## [1.8.1] - 2025-02-07